# src/routes/private-routes.js
import express from "express";
import authController from "../controllers/auth/login-controllers.js"
import taskController from "../controllers/task/taskControllers.js"
import authMiddleware from "../middleware/auth-middleware.js"
import autorizateMiddleware from "../middleware/autorization-middleware.js"
import { logMessage } from "../utils/log-generator.js";

import ordersRoutes from "./orders/orders-routes.js"
import creditRoutes from "./credit/credit-routes.js"
import paymentsRoutes from "./payments/payments-routes.js"
import pricesRoutes from "./prices/prices-routes.js"
import productsRoutes from "./products/products-routes.js"
import customersRoutes from "./customers/customers-routes.js"
import availabilityRoutes from "./availability/availability-routes.js"

const router = express.Router();

router.use(authMiddleware)	// Middleware de autenticação

router.use('/orders',autorizateMiddleware(['admin','agrega']),ordersRoutes);
router.use('/credit',autorizateMiddleware(['admin','agrega']),creditRoutes);
router.use('/payments',autorizateMiddleware(['admin','agrega']),paymentsRoutes);
router.use('/prices',autorizateMiddleware(['admin','agrega']),pricesRoutes);
router.use('/customers',autorizateMiddleware(['admin','agrega']),customersRoutes);
router.use('/products',autorizateMiddleware(['admin','agrega']),productsRoutes);
router.use('/availability',autorizateMiddleware(['admin','agrega']),availabilityRoutes);


router.post('/newUser',autorizateMiddleware(['admin']), async(req, res) => {
	
	try {
		const usersAdded = await authController.newUser(req.body) 
		if (usersAdded < 1)
			throw new Error("User include failed ");
		
		logMessage(`User ${req.body.login} added successful`)
		res.success(null,'user added successful.')
		
	} catch (error) {
		
		res.error(500,error,'User adding failed.');
	}
	
});

router.post('/newTaskAgrega', async(req, res) => {
	
	
	try {
		
		if(!taskController.newTask('agrega',req.body))
			throw new Error("task adding failed");
		
		logMessage(`Agrega Task added successful`)
		res.success(null, 'task added successful.')
		
		
	} catch (error) {
		res.error(500, error, 'Failure on add Agregga task')
	}
});

router.post('/newTaskSimulador', async (req, res) => {
	
	try {
		
		await taskController.newTask('simulador',req.body)
		logMessage(`Simulador Task added successful`)

		res.success(null, 'task added successful.')


	} catch (error) {
		res.error(500, error, 'Failure on add Simulador task')
	}
});


export default router;