# src/swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "Conecta by CCAB",
		"version": "1.0.0",
		"description": "API de integração com o ERP da CCAB",
		"contact": {
			"name": "CCAB",
			"email": "suporte@ccab-agro.com.br"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Development Local server "
		},
		{
			"url": "https://api.ccab-agro.net",
			"description": "Production server"
		}
	],
	"paths": {
		"/orders": {
			"get": {
				"summary": "Get all orders",
				"description": "Get the sales orders from the ERP.",
				"tags": [
					"Orders"
				],
				"parameters": [
					{
						"name": "pageSize",
						"in": "query",
						"description": "The number of orders to return",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"name": "pageNumber",
						"in": "query",
						"description": "Control of pagination on search by orders to return",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"name": "orderNumber",
						"in": "query",
						"description": "Filter orders by ID",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "client",
						"in": "query",
						"description": "Filter orders by client ID",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "emitDate",
						"in": "query",
						"description": "Filter orders by create data. (Format: YYYYMMDD)",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "priceTable",
						"in": "query",
						"description": "Filter orders by price table.",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "A list of orders according to the filters parameters",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OrderNoItems"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Send new order",
				"description": "Send to ERP the sales orders and financial information about a negotiation.",
				"tags": [
					"Orders"
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"order": {
										"$ref": "#/components/schemas/NewOrderData"
									},
									"payment": {
										"$ref": "#/components/schemas/Payment"
									}
								},
								"required": [
									"order",
									"payment"
								]
							}
						}
					}
				}
			}
		},
...
		"/credit": {
			"get": {
				"summary": "Get credit limits",
				"description": "Get credit limits of all the economic groups",
				"tags": [
					"Credit"
				],
				"parameters": [
					{
						"name": "pageSize",
						"in": "query",
						"description": "The number of registries to return",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"name": "pageNumber",
						"in": "query",
						"description": "Control of pagination on search",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"name": "group",
						"in": "query",
						"description": "Filter ordelimit credits  by ID of economic group",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "crop",
						"in": "query",
						"description": "Filter a limit of a specific crop. (Format: YYYY/YYYY). Default is the current crop.",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "A list of limit credit according to the filters parameters",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/CreditLimit"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/prices": {
			"get": {
				"summary": "Get prices tables without items",
				"description": "Get the price tables actives on the ERP without items.",
				"tags": [
					"Price Tables"
				],
				"parameters": [
					{
						"name": "pageSize",
						"in": "query",
						"description": "The number of registries to return",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"name": "pageNumber",
						"in": "query",
						"description": "Control of pagination on search",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"name": "crop",
						"in": "query",
						"description": "Filter by a specific crop. Format: AAAA/AAAA",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "A list of price tables according to the filters parameters",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/PriceTableNoItems"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/customers": {
			"get": {
				"summary": "Get a list of customers",
				"description": "Get a list of active customers from the ERP.",
				"tags": [
					"Customers"
				],
				"parameters": [
					{
						"name": "pageSize",
						"in": "query",
						"description": "The number of registries to return",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"name": "pageNumber",
						"in": "query",
						"description": "Control of pagination on search",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"name": "state",
						"in": "query",
						"description": "Filter by a specific state of the federation.",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "name",
						"in": "query",
						"description": "Search for customers whose name contains the text entered.",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "economicGroup",
						"in": "query",
						"description": "Filter by a specific economic group.",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "A list of customers according to the filters parameters",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Customer"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/products":{
			"get": {
				"summary": "Get list of products ",
				"description": "Get the list of all comercial products according the filter informed.",
				"tags": [
					"Products"
				],
				"parameters": [
					{
						"name": "alias",
						"in": "query",
						"description": "The ID of the chemical alias to search for.",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "package",
						"in": "query",
						"description": "The ID of the package to search for. Example: Glifosate on package 20l.",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "group",
						"in": "query",
						"description": "The ID of the product group to search for like insecticides, herbicides, etc.",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "An array with the products according the filter.",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Product"
									}
								}
								
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"OrderNoItems": {
				"type": "object",
				"properties": {
					"orderNumber": {
						"type": "string"
					},
					"emitDate": {
						"type": "string"
					},
					"customerId": {
						"type": "string"
					},
					"customerStore": {
						"type": "string"
					}
				}
			},
			"NewOrderData": {
				"type": "object",
				"properties": {
					"customerId": {
						"type": "string",
						"description": "Customer ID"
					},
					"customerStore": {
						"type": "string",
						"description": "Customer Store"
					},
					"priceTable": {
						"type": "string",
						"description": "Price table id"
					},
					"paymentCondition": {
						"type": "string",
						"description": "Payment Condition id"
					},
					"dueDate": {
						"type": "string",
						"description": "Due Date. Format: [AAAAMMDD]"
					},
					"items": {
						"type": "array",
						"items": {
							"type": "object",
							"properties": {
								"item": {
									"type": "integer",
									"description": "Item number"
								},
								"product": {
									"type": "string",
									"description": "Product ID"
								},
								"amount": {
									"type": "integer",
									"description": "Amount of product"
								},
								"price": {
									"type": "number",
									"description": "value negotiated for the product"
								},
								"deliveryDate": {
									"type": "string",
									"description": "Delivery date negotiated for this item. Format: [AAAAMMDD]"
								}
							}
						}
					}
				}
			},
			"CreditLimit": {
				"type": "object",
				"properties": {
					"group": {
						"type": "string"
					},
					"currency": {
						"type": "string"
					},
					"actualLimit": {
						"type": "number"
					}
				}
			},
			"Payment": {
				"type": "object",
				"properties": {
					"documentNumber":{
						"type": "string"
					},
					"parcel":{
						"type": "string"
					},
					"type":{
						"type": "string"
					},
					"clientId":{
						"type": "string"
					},
					"clientStore":{
						"type": "string"
					},
					"issueDate":{
						"type": "string"
					},
					"dueDate":{
						"type": "string"
					},
					"balance":{
						"type": "number"
					},
					"currency":{
						"type": "string"
					},
					"description":{
						"type": "string"
					}
				}
			},
			"PriceTableNoItems": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"crop": {
						"type": "string"
					}
				}
			},
...
			"Customer": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"store": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"idDocument": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"farm": {
						"type": "string"
					},
					"economicGroup": {
						"type": "string"
					}
				}
			},
...
			"Product": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"alias": {
						"type": "string"
					},
					"aliasDescription": {
						"type": "string"
					},
					"package": {
						"type": "string"
					},
					"conversionFactor": {
						"type": "number"
					},
					"leatTime": {
						"type": "number"
					},
					"group": {
						"type": "string"
					},
					"groupDescription": {
						"type": "string"
					}
				}
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	]
}